Fix stop all stations button
Zoek Mach - Cancel the running program and stop all stations, when a button is pressed. This is exactly the same that can be done by pressing the “Stop all stations” button on the OSPy homepage. But when using the button control setting “Stop all stations”, only the running station will be skipped and if a program is active, immediately the next station will start.